Move name registry to codegen
