+++ generators/swift/base/src/project/index.ts
@@ -1,3 +1,2 @@
-export * from "./name-registry";
 export * from "./SwiftFile";
 export * from "./SwiftProject";
