+++ generators/swift/codegen/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./custom-config";
 export * from "./helpers";
+export * from "./name-registry";
 export * as swift from "./swift";
 export * from "./syntax";
