+++ generators/swift/codegen/src/name-registry/index.ts
