+++ generators/swift/codegen/src/name-registry/requests-namespace.ts
