+++ generators/swift/codegen/src/name-registry/source-module-namespace.ts
