+++ generators/swift/codegen/src/name-registry/test-module-namespace.ts
