+++ generators/swift/codegen/src/symbol/index.ts
@@ -1,2 +1,2 @@
-export { type FoundationTypeSymbolName, type SwiftTypeSymbolName, Symbol } from "./symbol";
+export { type AsIsSymbolName, type FoundationTypeSymbolName, type SwiftTypeSymbolName, Symbol } from "./symbol";
 export { SymbolRegistry, type TypeSymbolShape } from "./symbol-registry";
